requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Mail>=0.9.1
Flask-Migrate>=3.1.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
plotly>=5.3.1
pandas>=1.3.3
numpy>=1.21.2
openai>=1.0.0
pymongo>=3.12.0
sendgrid>=6.9.0
supabase>=1.0.0
Werkzeug>=2.0.0
gunicorn>=20.1.0
certifi>=2023.0.0